cart_service/requirements.txt
Django==5.1.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-redis==5.2.0
redis==4.5.5
gunicorn==21.2.0
python-dotenv==1.0.0